base: added systemd.
This is less than ideal, as we will ship BOTH sysv-rc and systemd.
Unfortunately due to the current systemd situation in Debian (unable to reliably replace sysv-rc) and in
Debian Live (full systemd support in 4.x, which is in experimental), we are going to ship both.

The Live system will boot with sysv-rc.
A trivial change to linstaller's bootloader.inst module will be needed to make the installed system boot
with systemd.

NOTE: This needs to be changed with systemd-sysv AS SOON AS POSSIBLE.